+++ NEWS.md
@@ -1,3 +1,12 @@
+CHANGES IN VERSION 1.1.6
+-------------------------
+
+* Updated installation information in README
+* "eruption" gained argument "labelCutoffs" to add labels to p value and fold change cutoff lines
+* "ppiPlotNetwork" checks for argument "labelColumn" when "label=TRUE"
+* Fixed some tests
+
+
 CHANGES IN VERSION 0.99.380
 ----------------------------